Declare create-grid and product and a for loop
CreateGrid needs to be declared because it will contain the most
important code. Product needs to be declared so that its value can be
used to create the correct number of divs that will make up the grid.
The for loop is necessary for automating the process of making the
divs.