Prevent cancelling prompt from erasing gridDivs
The code allowing the user to break the endless loop by hitting esc or
cancel when prompted was inside of the if statement that also removes
the previous grid. This caused a bug where everything with the class
gridDiv would be erased even if the user just cancelled the prompt.
This was fixed by giving the code its own else if statement that
only breaks and does not remove anything.