Declare functions getRandomRGB and getRandomNumber
Both functions will be necessary for randomizing the color of all
gridDivs. GetRandomNumber will generate three random numbers that
will be assigned to the red, green, and blue parameters of
getRandomRGB. Then red, green, and blue will be returned and used
as values for backgroundColor in the form of variables.