Make create-grid automatically create containers
Fixes an issue causing the grid to not display correctly. For example,
instead of having a height and width of 4, it may have a height of 8
and a width of 2. While this would result in the same product, it was
not ideal.

This should also fix an issue that makes some gridSquares different
sizes.