Add height and width to outermost-container
Specifies how much space outermost-container is allowed to take up.
Failing to do this will make it not take up any space, preventing
the grid inside of outer-container from appearing on screen. The
height is set to exactly 96% because this stops the grid from expanding
into an area that would require the user to scroll.